now handling success results from lib3h
